Create Clarification
Hi All
In this article, it's mentioned that adding distribution groups in nesting scenarios is not supported , however there is another article mentioned that we can add a group 
(Membership
Use this tab to add or remove group members. Group owners don't need to be members of the group.
To add members, click Add Add icon.. In the dialog that appears, find, and select a recipient or group,)
https://docs.microsoft.com/en-us/exchange/recipients-in-exchange-online/manage-distribution-groups/manage-distribution-groups
I'd like to know if adding distribution groups in nesting scenarios is supported?